Restructure the code-complexity tool into pipeline stages
The tool was 27 flat entries in one directory with four doc files scattered
among them, and CONTEXT.md served four different readers in one linear scroll.

Group the sources into a directory per pipeline stage: select/ (scope, diff),
measure/ (complexity, jsx, crap), coverage/ (read, generate), render/ (report,
github). index.ts, config.ts and analyze.ts stay at the root. Rename
coverage.ts and generate-coverage.ts to coverage/read.ts and coverage/generate.ts
so the paths stop stuttering. Specs move with their sources; no logic changes.

Split the docs by reader question and collect them under docs/: MODEL.md
(renamed from SCORING.md, now also carrying the scope rules and the jsx/behavior
classification), RUNNING.md (new: the two selection axes, generation scope, the
report, CI wiring) and CALIBRATION.md. adr/ stays at the module root, per
.agents/rules/adr.md. CONTEXT.md drops to an entry point: a routing table, the
gate-failed triage, and a file map grouped by stage.

Reframe the tool as a code-quality gate rather than a CI gate, and say plainly
that select/, measure/ and coverage/ never see a cap: the thresholds enter once,
in analyze.ts, and everything downstream reads the verdict it stamps.
