+++ tools/code-complexity/adr/0001-optional-chaining-excluded-from-complexity.md
@@ -44,9 +44,9 @@ settled at calibration.
 - The gate is a **conscious divergence** from ESLint's cyclomatic definition. A future reader — or a
   PR trying to align the tool with a standard complexity implementation — will see the gap and may
   try to "fix" it back; this record is what says not to. The counting convention and this caveat
-  live in `../SCORING.md`.
+  live in `../docs/MODEL.md`.
 - The gate's calibrated thresholds (`../config.ts`) were set from the distribution *with `?.` already
   dropped*. Re-introducing `?.` counting would silently make the caps far more aggressive; the two
   decisions are coupled, and changing one demands recalibrating the other.
-- Reversing is a one-case change in `countsAsBranch` (`../complexity.ts`) plus its spec — cheap in
+- Reversing is a one-case change in `countsAsBranch` (`../measure/complexity.ts`) plus its spec — cheap in
   code, but it invalidates the calibration, so it is an ADR-level decision, not a tweak.
