+++ tools/code-complexity/render/github.ts
@@ -2,8 +2,8 @@
 // offender, plus the full report table written to the job summary. Pure formatters here; the
 // side effects (stdout + $GITHUB_STEP_SUMMARY) live in index.ts, guarded by $GITHUB_ACTIONS.
 
-import type { Contribution } from './complexity';
-import { flatteningSaving, isNestingStructureReason } from './complexity';
+import type { Contribution } from '../measure/complexity';
+import { flatteningSaving, isNestingStructureReason } from '../measure/complexity';
 import type { ReportRow, Thresholds } from './report';
 import { isOffender, maxCognitiveFor } from './report';
