+++ tools/code-complexity/render/report.ts
@@ -1,9 +1,9 @@
 // Renders the results table: every checked function with its cognitive complexity, coverage%, CRAP,
 // and which threshold (if any) it broke, sorted by CRAP descending with offenders flagged.
 
-import type { Contribution } from './complexity';
+import type { Contribution } from '../measure/complexity';
 
-// A function's class, decided by whether JSX appears directly in its own body (./complexity.ts).
+// A function's class, decided by whether JSX appears directly in its own body (../measure/complexity.ts).
 export type FunctionKind = 'jsx' | 'behavior';
 
 export interface ReportRow {
