+++ tools/code-complexity/select/scope.spec.ts
