+++ tools/code-complexity/select/scope.ts
